Port over to using hyper instead of rust-http, because it is deprecated
rust-http is a deprecated library, hyper is one the primary replacements.

hyper is slightly less stable than rust-http, but it is under more
active development and provides a much nicer API than rust-http.

Most of hyper's most important APIs, such as basic header setting,
status, and writing to the body of a response are already basically
stabilized.